wrap atproto search params usage in suspense boundary
